requirements.txt
# --- Core Framework ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# --- Google Gen AI (NEW SDK — replaces deprecated google-generativeai) ---
google-genai>=1.0.0

# --- Google Cloud Services ---
google-cloud-firestore>=2.19.0
google-api-python-client>=2.150.0
google-auth>=2.35.0

# --- HTTP Client (async, used for Maps Places API v2) ---
httpx>=0.27.0

# --- Auth ---
python-jose[cryptography]>=3.3.0

# --- File Handling ---
python-multipart>=0.0.12
Pillow>=10.4.0

# --- Templating (weekly report) ---
jinja2>=3.1.4

# --- Testing ---
pytest>=8.3.0
pytest-asyncio>=0.24.0
